+++ src/main/java/fll/web/admin/PerformanceSchedule.java
@@ -46,7 +46,7 @@ public class PerformanceSchedule extends BaseFLLServlet {
       throws IOException, ServletException {
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.requireRoles(request, response, session, Set.of(UserRole.ADMIN), false)) {
+    if (!auth.requireRoles(request, response, session, Set.of(UserRole.PUBLIC), false)) {
       return;
     }