+++ src/main/java/fll/web/admin/PerformanceSheets.java
@@ -48,7 +48,7 @@ public class PerformanceSheets extends BaseFLLServlet {
       throws IOException, ServletException {
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.requireRoles(request, response, session, Set.of(UserRole.ADMIN), false)) {
+    if (!auth.requireRoles(request, response, session, Set.of(UserRole.PUBLIC), false)) {
       return;
     }